Added the Weapon Interface to bind them together. Additionally, splitted the Weapon class into
