+++ src/main/java/weapon/LifestealWeapon.java
@@ -1,0 +1,4 @@
+package weapon;
+
+public abstract class LifestealWeapon implements Weapon {
+}
